Move ZhaRadioManager to a separate module
